Store the app type for DBR_CTRL requests then post complete GDDs (fixes #11)
Store the 'highest' GDD event app type if its DBR equivalent is between
DBR_CTRL_SHORT and DBR_CTRL_DOUBLE. Use this app type to construct a
temporary GDD in vcPostEvent() and do a copy_state() to fill that with
actual data then post this GDD instead of event_data.

This should fix #11.