Fix copying of enum labels.
Enum labels where not copied correctly when they changed for an active
monitor. This means that a monitor of type DBR_GR_ENUM or DBR_CTRL_ENUM
would receive the old labels, even though the gateway had actually
received and processed updated labels from the CA server.

This commit addresses this issue by ensuring that all labels are copied
to the target GDD when processing updates.